+++ envoy-control-core/src/test/java/pl/allegro/tech/servicemesh/envoycontrol/v3/SimpleCacheWithMissingEndpointsTest.java
@@ -61,7 +61,9 @@ public class SimpleCacheWithMissingEndpointsTest extends SimpleCacheTest {
                         .build()),
                 Collections.emptySet(),
                 responseTracker,
-                false);
+                false,
+                false
+            );
 
         assertThatWatchReceivesSnapshotWithMissingResources(new WatchAndTracker(watch, responseTracker), SNAPSHOT_WITH_MISSING_RESOURCES);
     }