+++ envoy-control-core/src/test/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/resource/listeners/filters/rbac/RBACFilterFactoryTest.kt
@@ -225,7 +225,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     @Test
     fun `should generate RBAC rules for incoming permissions with log unlisted clients and endpoints`() {
         // given
-        val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
+        val policyName = "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
             "clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], " +
             "unlistedClientsPolicy=LOG, oauth=null)"
         val expectedShadowRules = expectedSimpleEndpointPermissionsJson(policyName)
@@ -238,6 +238,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
             permissionsEnabled = true,
             endpoints = listOf(
                 IncomingEndpoint(
+                    emptySet(),
                 "/example",
                     PathMatchingType.PATH,
                     setOf("GET", "POST"),
@@ -258,7 +259,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     @Test
     fun `should generate RBAC rules for incoming permissions with block unlisted endpoints and log clients`() {
         // given
-        val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
+        val policyName = "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
             "clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], " +
             "unlistedClientsPolicy=LOG, oauth=null)"
         val expectedShadowRules = expectedSimpleEndpointPermissionsJson(policyName)
@@ -271,6 +272,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                 permissionsEnabled = true,
                 endpoints = listOf(
                         IncomingEndpoint(
+                                emptySet(),
                                 "/example",
                                 PathMatchingType.PATH,
                                 setOf("GET", "POST"),
@@ -291,7 +293,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     @Test
     fun `should generate RBAC rules for incoming permissions with log unlisted endpoints and block clients`() {
         // given
-        val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
+        val policyName = "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
             "clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], " +
             "unlistedClientsPolicy=BLOCKANDLOG, oauth=null)"
         val expectedShadowRules = expectedSimpleEndpointPermissionsJson(policyName)
@@ -304,6 +306,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                 permissionsEnabled = true,
                 endpoints = listOf(
                         IncomingEndpoint(
+                                emptySet(),
                                 "/example",
                                 PathMatchingType.PATH,
                                 setOf("GET", "POST"),
@@ -324,12 +327,13 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     @Test
     fun `should generate RBAC rules for incoming permissions with roles`() {
         // given
-        val policyName = "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
+        val policyName = "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], " +
             "clients=[ClientWithSelector(name=role-1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)"
         val expectedRbacBuilder = getRBACFilter(expectedSimpleEndpointPermissionsJson(policyName))
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
@@ -351,6 +355,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
@@ -383,11 +388,13 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
                         setOf(ClientWithSelector.create("client1"))
                 ), IncomingEndpoint(
+                        emptySet(),
                         "/example2",
                         PathMatchingType.PATH,
                         setOf("POST"),
@@ -408,11 +415,13 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
                         setOf(ClientWithSelector.create("role-1"))
                 ), IncomingEndpoint(
+                        emptySet(),
                         "/example2",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
@@ -436,11 +445,13 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
                         setOf(ClientWithSelector.create("client2"), ClientWithSelector.create("role-1"))
                 ), IncomingEndpoint(
+                        emptySet(),
                         "/example2",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
@@ -465,6 +476,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET", "POST"),
@@ -485,6 +497,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf()
@@ -506,7 +519,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val expectedActual = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[], clients=[], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[], clients=[], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [{
                 "and_rules": {
                   "rules": [ ${pathRule("/example")} ]
@@ -601,6 +614,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
@@ -622,6 +636,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
@@ -646,6 +661,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
@@ -670,6 +686,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
@@ -694,6 +711,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
                 permissionsEnabled = true,
                 endpoints = listOf(IncomingEndpoint(
+                        emptySet(),
                         "/example",
                         PathMatchingType.PATH,
                         setOf("GET"),
@@ -722,6 +740,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
             permissionsEnabled = true,
             endpoints = listOf(IncomingEndpoint(
+                emptySet(),
                 "/example",
                 PathMatchingType.PATH,
                 setOf("GET"),
@@ -745,6 +764,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
             permissionsEnabled = true,
             endpoints = listOf(IncomingEndpoint(
+                emptySet(),
                 "/default",
                 PathMatchingType.PATH,
                 setOf("GET"),
@@ -768,6 +788,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         val incomingPermission = Incoming(
             permissionsEnabled = true,
             endpoints = listOf(IncomingEndpoint(
+                emptySet(),
                 "/custom",
                 PathMatchingType.PATH,
                 setOf("GET"),
@@ -788,7 +809,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedEndpointPermissionsWithDifferentRulesForDifferentClientsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -808,7 +829,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
                 ${originalAndAuthenticatedPrincipal("client1")}
               ]
             },
-            "IncomingEndpoint(path=/example2, pathMatchingType=PATH, methods=[POST], clients=[ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example2, pathMatchingType=PATH, methods=[POST], clients=[ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -835,7 +856,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpFromDiscoveryWithSelectorAuthPermissionsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=selector, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=selector, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -870,7 +891,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpWithSelectorAuthPermissionsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client2, selector=selector, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client2, selector=selector, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -905,7 +926,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpWithStaticRangeAndSelectorAuthPermissionsAndRolesJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=role1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=role1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -949,7 +970,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpAuthPermissionsJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1008,7 +1029,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
         {
           "policies": {
            """ /* notice that duplicated clients occurs only once here */ + """
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client1, selector=selector, negated=false), ClientWithSelector(name=client1-duplicated, selector=selector, negated=false), ClientWithSelector(name=client1-duplicated, selector=null, negated=false), ClientWithSelector(name=role-1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client1, selector=selector, negated=false), ClientWithSelector(name=client1-duplicated, selector=selector, negated=false), ClientWithSelector(name=client1-duplicated, selector=null, negated=false), ClientWithSelector(name=role-1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1228,7 +1249,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpAuthWithStaticRangeJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1255,7 +1276,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedSourceIpAuthWithStaticRangeAndSourceIpJson = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1292,7 +1313,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private val expectedEndpointPermissionsLogUnlistedEndpointsAndBlockUnlistedClients = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET, POST], clients=[ClientWithSelector(name=client1, selector=null, negated=false), ClientWithSelector(name=client2, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                     "and_rules": {
@@ -1348,7 +1369,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private fun expectedPoliciesForAllowedClient(principals: String) = """
         {
           "policies": {
-            "IncomingEndpoint(path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=/example, pathMatchingType=PATH, methods=[GET], clients=[ClientWithSelector(name=client1, selector=null, negated=false)], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {
@@ -1373,7 +1394,7 @@ internal class RBACFilterFactoryTest : RBACFilterFactoryTestUtils {
     private fun expectedPoliciesForDefaultAndCustomLists(clients: List<String>, principals: List<String>, path: String) = """
         {
           "policies": {
-            "IncomingEndpoint(path=$path, pathMatchingType=PATH, methods=[GET], clients=[${clients.joinToString(", ") { "ClientWithSelector(name=$it, selector=null, negated=false)" }}], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
+            "IncomingEndpoint(paths=[], path=$path, pathMatchingType=PATH, methods=[GET], clients=[${clients.joinToString(", ") { "ClientWithSelector(name=$it, selector=null, negated=false)" }}], unlistedClientsPolicy=BLOCKANDLOG, oauth=null)": {
               "permissions": [
                 {
                   "and_rules": {